xiaomi-aiot-edge-security/requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pyserial==3.5
jsonschema==4.19.1
urllib3==2.0.7
pydantic==2.5.2
Flask==2.2.5
Werkzeug==2.2.3

edgexfoundry-device-sdk-python==3.1.0
ThingsBoard==3.9.11
tb-mqtt-client==1.5.1
paho-mqtt==1.6.1

pycryptodome==3.19.0
cryptography==41.0.5
pyopenssl==23.3.0
bcrypt==4.0.1
passlib==1.7.4

pandas==2.1.1
numpy==1.26.1
matplotlib==3.8.1
seaborn==0.13.0
scipy==1.11.3
scikit-learn==1.3.2
reportlab==4.0.7

flask==2.3.3
flask-socketio==5.3.6
flask-wtf==1.2.1
flask-login==0.6.3
werkzeug==2.3.7
jinja2==3.1.2
wtforms==3.1.1

simpy==4.0.2
networkx==3.2.1

pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

tqdm==4.66.1
colorama==0.4.6
click==8.1.7
